requirements.txt
anthropic>=0.49.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2024.1
lxml>=5.1.0
twilio>=9.0.0
gunicorn>=21.2.0
